+++ web-test-runner.config.js
@@ -22,13 +22,20 @@ export default {
   nodeResolve: true,
   browsers: [
     playwrightLauncher({product: 'chromium'}),
-    playwrightLauncher({product: 'firefox'}),
+    // TODO(aomarks) Firefix is flaky, with service worker 404s. Probably needs
+    // to be addressed as part of
+    // https://github.com/PolymerLabs/code-sample-editor/issues/39
+    // playwrightLauncher({product: 'firefox'}),
     playwrightLauncher({product: 'webkit'}),
   ],
+  browserStartTimeout: 30000, // default 30000
+  testsStartTimeout: 20000, // default 10000
+  testsFinishTimeout: 60000, // default 20000
   testFramework: {
     // https://mochajs.org/api/mocha
     config: {
       ui: 'tdd',
+      timeout: '10000', // default 2000
     },
   },
 };